third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/xmlhttprequest-sync-default-feature-policy.sub-expected.txt

This is a testharness.js-based test.
PASS Default "sync-xhr" feature policy ["*"] allows the top-level document.
PASS Default "sync-xhr" feature policy ["*"] allows same-origin iframes.
PASS Default "sync-xhr" feature policy ["*"] allows cross-origin iframes.
FAIL Feature policy "sync-xhr" can be disabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#InvalidAccessError: Failed to execute 'open' on 'XMLHttpRequest': Synchronous requests are disabled by Feature Policy." but got "#OK"
Harness: the test ran to completion.
